requirements.txt
altgraph==0.16.1
bottlenose==1.1.8
future==0.17.1
lxml==4.3.4
macholib==1.11
pefile==2019.4.18
pyinstaller==3.4
pysimplegui==4.0.0
python-amazon-simple-product-api==2.2.11
python-dateutil==2.8.0
pywin32-ctypes==0.2.0
six==1.12.0